new imagehandler

+++ utils/imageHandler.py
@@ -2,7 +2,7 @@ import os
 import json
 import time
 import json
-from PIL import Image
+from PIL import Image, ImageOps
 from subprocess import Popen
 
 
@@ -15,7 +15,7 @@ def handle():
 
     # sort the directory by when the file was last accessed (i.e. when I put
     # the file in the folder)
-    files.sort(key=os.path.getatime, reverse=1)
+    files.sort(key=os.path.getmtime, reverse=1)
 
     # open the javascript file
     scriptFile = open("../script.js", "r")
@@ -25,13 +25,14 @@ def handle():
     # Cut out only the wallpaper array from the script file
     walls = fileStr[(fileStr.find("[") + 1):(fileStr.find("]"))]
 
-    if(walls != ""):
+    print(walls)
+
+    if (walls != "" and len(walls) > 0):
         walls = walls.replace("\n", "")
         walls = "[" + walls + "]"
-        # print(walls)
         walls = json.loads(walls)
         # get when the last wallpaper was loaded
-        prevLoadTime = walls[0].get("added")
+        prevLoadTime = float(walls[0].get("added")) if len(walls) > 0 else 0
         print("prevloadTime: " + str(prevLoadTime))
     else:
         prevLoadTime = 0
@@ -43,29 +44,43 @@ def handle():
 
     for file in files:
         print("The most recent file is: " + file +
-              " with load time: " + str(os.path.getatime("../walls/" + file)))
-        if(os.path.getatime("../walls/" + file) > prevLoadTime):
-            newWalls.append(
-                {"filename": file, "added": os.path.getatime("../walls/" + file)})
-            print("found a new wallpaper: " + file + ", generating thumbnail!")
-            img = Image.open("../walls/" + file)
-            if img.mode != "RGB":
-                img = img.convert("RGB")
-            img.thumbnail(size)
-            img.save("../thumbs/" + file, "JPEG", quality=50)
-            print("saving: " + file)
-        else:
-            break
+              " with load time: " + str(os.path.getmtime("../walls/" + file)))
+        # if(os.path.getmtime("../walls/" + file) > prevLoadTime):
+        #     newWalls.append(
+        #         {"filename": file, "added": os.path.getmtime("../walls/" + file)})
+        #     print("found a new wallpaper: " + file + ", generating thumbnail!")
+        #     print("this has mtime: " + os.path.getmtime("../walls/" + file))
+        #     img = Image.open("../walls/" + file)
+        #     if img.mode != "RGB":
+        #         img = img.convert("RGB")
+        #     img.thumbnail(size)
+        #     img.save("../thumbs/" + file, "JPEG", quality=50)
+        #     print("saving: " + file)
+        # else:
+        #     break
+        newWalls.append(
+            {"filename": file, "added": os.path.getmtime("../walls/" + file)})
+        print("found a new wallpaper: " + file + ", generating thumbnail!")
+        img = Image.open("../walls/" + file)
+        if img.mode != "RGB":
+            img = img.convert("RGB")
+        img.thumbnail(size)
+        img = ImageOps.exif_transpose(img)
+        img.save("../thumbs/" + file, "JPEG", quality=50)
+        print("saving: " + file)
 
     for wall in walls:
         wallFile = wall.get("filename").strip()
         print("wall: " + wallFile)
-        if(wall != "" and os.path.isfile("../walls/" + wallFile)):
+        if (wall != "" and os.path.isfile("../walls/" + wallFile)):
             newWalls.append({"filename": wallFile, "added": wall.get("added")})
-        elif(wallFile != "" and (not os.path.isfile('../walls/' + wallFile))):
-            print(
-                wallFile + " has been removed as a thumbnail and from the website.")
-            os.remove('../thumbs/' + wallFile)
+        elif (wallFile != "" and (not os.path.isfile('../walls/' + wallFile))):
+            if (os.path.isfile('../thumbs/' + wallFile)):
+                print(
+                    wallFile + " has been removed as a thumbnail and from the website.")
+                os.remove('../thumbs/' + wallFile)
+            else:
+                print(wallFile + " has already been removed as a thumbnail")
 
     print(newWalls)
 
@@ -80,6 +95,7 @@ def handle():
 
     os.chdir("../utils")
     # Popen("gitHelp.bat", cwd=os.getcwd())
+
 
 if __name__ == "__main__":
     handle()
